requirementos.txt
aiohttp==3.8.3
aiosignal==1.2.0
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
billiard==3.6.4.0
Brotli==1.0.9
cachetools==5.2.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==4.0.0
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
contextlib2==21.6.0
cryptography==38.0.1
daphne==4.0.0
db-dtypes==1.0.4
decorator==5.1.1
Deprecated==1.2.13
dj-database-url==1.0.0
Django==3.2.16
django-cors-headers==3.13.0
django-heroku==0.3.1
django-phonenumber-field==7.0.0
django-redis==5.2.0
django-sql-explorer==2.5
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
exceptiongroup==1.0.0rc9
Faker==12.0.1
frozenlist==1.3.1
fsspec==2022.10.0
gcsfs==2022.10.0
google-api-core==2.10.2
google-auth==2.13.0
google-auth-oauthlib==0.7.0
google-cloud-bigquery==3.3.5
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-cloud-datastore==1.15.5
google-cloud-ndb==1.11.1
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.56.4
grpcio==1.50.0
grpcio-status==1.48.2
gunicorn==20.1.0
h2==4.1.0
hiredis==2.0.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
iniconfig==1.1.1
kombu==5.2.4
lxml==4.9.1
mixer==7.2.2
multidict==6.0.2
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
pandas==1.5.1
phone-iso3166==0.4.1
phonenumberslite==8.12.57
Pillow==9.2.0
pluggy==1.0.0
priority==1.3.0
prompt-toolkit==3.0.31
proto-plus==1.22.1
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.5
psycopg2-binary==2.9.5
py==1.11.0
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyfcm==1.5.4
PyGithub==1.56
PyJWT==2.6.0
pymemcache==4.0.0
PyNaCl==1.5.0
pyOpenSSL==22.1.0
pyparsing==3.0.9
pytest==7.2.0
pytest-django==4.5.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-slugify==6.1.2
pytz==2022.5
PyYAML==6.0
redis==4.3.4
requests==2.28.1
requests-oauthlib==1.3.1
rollbar==0.16.3
rsa==4.9
schema==0.7.5
serpy==0.3.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
timeago==1.0.16
tomli==2.0.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
wrapt==1.14.1
XlsxWriter==3.0.3
yarl==1.8.1
zope.interface==5.5.0
